+++ src/main/java/com/gdschongik/gdsc/domain/membership/dao/MembershipRepository.java
@@ -3,12 +3,13 @@ package com.gdschongik.gdsc.domain.membership.dao;
 import com.gdschongik.gdsc.domain.member.domain.Member;
 import com.gdschongik.gdsc.domain.membership.domain.Membership;
 import com.gdschongik.gdsc.domain.recruitment.domain.RecruitmentRound;
+import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MembershipRepository extends JpaRepository<Membership, Long> {
 
-    Optional<Membership> findByMember(Member member);
+    List<Membership> findAllByMember(Member member);
 
     Optional<Membership> findByMemberAndRecruitmentRound(Member member, RecruitmentRound recruitmentRound);
 }
