+++ src/test/java/com/gdschongik/gdsc/domain/membership/application/MembershipServiceTest.java
@@ -37,6 +37,38 @@ public class MembershipServiceTest extends IntegrationTest {
                     .isInstanceOf(CustomException.class)
                     .hasMessage(RECRUITMENT_ROUND_NOT_FOUND.getMessage());
         }
+
+        @Test
+        void 해당_학기에_이미_Membership을_발급받았다면_실패한다() {
+            // given
+            Member member = createMember();
+            logoutAndReloginAs(1L, ASSOCIATE);
+            RecruitmentRound recruitmentRound = createRecruitmentRound();
+            Membership membership = createMembership(member, recruitmentRound);
+
+            // when
+            membership.verifyPaymentStatus();
+            membershipRepository.save(membership);
+
+            // then
+            assertThatThrownBy(() -> membershipService.submitMembership(recruitmentRound.getId()))
+                    .isInstanceOf(CustomException.class)
+                    .hasMessage(MEMBERSHIP_ALREADY_SATISFIED.getMessage());
+        }
+
+        @Test
+        void 해당_학기에_이미_Membership을_생성한_적이_있다면_실패한다() {
+            // given
+            Member member = createMember();
+            logoutAndReloginAs(1L, ASSOCIATE);
+            RecruitmentRound recruitmentRound = createRecruitmentRound();
+            createMembership(member, recruitmentRound);
+
+            // when & then
+            assertThatThrownBy(() -> membershipService.submitMembership(recruitmentRound.getId()))
+                    .isInstanceOf(CustomException.class)
+                    .hasMessage(MEMBERSHIP_ALREADY_SUBMITTED.getMessage());
+        }
     }
 
     @Test
