+++ src/test/java/com/gdschongik/gdsc/domain/membership/domain/MembershipValidatorTest.java
@@ -1,79 +1,33 @@
 package com.gdschongik.gdsc.domain.membership.domain;
 
-import static com.gdschongik.gdsc.domain.member.domain.MemberRole.*;
 import static com.gdschongik.gdsc.global.common.constant.RecruitmentConstant.*;
+import static com.gdschongik.gdsc.global.common.constant.SemesterConstant.*;
 import static com.gdschongik.gdsc.global.exception.ErrorCode.*;
 import static org.assertj.core.api.Assertions.*;
-import static org.mockito.Mockito.*;
 
-import com.gdschongik.gdsc.domain.member.domain.Member;
-import com.gdschongik.gdsc.domain.membership.dao.MembershipRepository;
+import com.gdschongik.gdsc.domain.recruitment.domain.Recruitment;
 import com.gdschongik.gdsc.domain.recruitment.domain.RecruitmentRound;
+import com.gdschongik.gdsc.domain.recruitment.domain.vo.Period;
 import com.gdschongik.gdsc.global.exception.CustomException;
-import com.gdschongik.gdsc.helper.IntegrationTest;
-import java.util.Optional;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 
-class MembershipValidatorTest extends IntegrationTest {
+class MembershipValidatorTest {
 
-    private MembershipValidator membershipValidator;
-    private MembershipRepository membershipRepository;
-
-    @BeforeEach
-    public void setUp() {
-        membershipRepository = Mockito.mock(MembershipRepository.class);
-        membershipValidator = new MembershipValidator(membershipRepository);
-    }
+    MembershipValidator membershipValidator = new MembershipValidator();
 
     @Nested
     class 멤버십_가입신청시 {
         @Test
-        void 해당_학기에_이미_Membership을_발급받았다면_실패한다() {
-            // given
-            Member member = createMember();
-            logoutAndReloginAs(1L, ASSOCIATE);
-            RecruitmentRound recruitmentRound = createRecruitmentRound();
-            Membership membership = createMembership(member, recruitmentRound);
-
-            when(membershipRepository.findByMember(member)).thenReturn(Optional.of(membership));
-
-            // when
-            membership.verifyPaymentStatus();
-
-            // when & then
-            assertThatThrownBy(() -> membershipValidator.validateMembershipSubmit(member, recruitmentRound))
-                    .isInstanceOf(CustomException.class)
-                    .hasMessage(MEMBERSHIP_ALREADY_SATISFIED.getMessage());
-        }
-
-        @Test
-        void 해당_Recruitment에_대해_Membership을_생성한_적이_있다면_실패한다() {
-            // given
-            Member member = createMember();
-            logoutAndReloginAs(1L, ASSOCIATE);
-            RecruitmentRound recruitmentRound = createRecruitmentRound();
-            Membership membership = createMembership(member, recruitmentRound);
-
-            when(membershipRepository.findByMember(member)).thenReturn(Optional.of(membership));
-
-            // when & then
-            assertThatThrownBy(() -> membershipValidator.validateMembershipSubmit(member, recruitmentRound))
-                    .isInstanceOf(CustomException.class)
-                    .hasMessage(MEMBERSHIP_ALREADY_SUBMITTED.getMessage());
-        }
-
-        @Test
         void 해당_RecruitmentRound의_모집기간이_아니라면_실패한다() {
             // given
-            Member member = createMember();
-            logoutAndReloginAs(1L, ASSOCIATE);
-            RecruitmentRound recruitmentRound = createRecruitmentRound();
+            Recruitment recruitment = Recruitment.createRecruitment(
+                    ACADEMIC_YEAR, SEMESTER_TYPE, FEE, Period.createPeriod(SEMESTER_START_DATE, SEMESTER_END_DATE));
+            RecruitmentRound recruitmentRound =
+                    RecruitmentRound.create(RECRUITMENT_NAME, START_DATE, END_DATE, recruitment, ROUND_TYPE);
 
             // when & then
-            assertThatThrownBy(() -> membershipValidator.validateMembershipSubmit(member, recruitmentRound))
+            assertThatThrownBy(() -> membershipValidator.validateMembershipSubmit(recruitmentRound))
                     .isInstanceOf(CustomException.class)
                     .hasMessage(RECRUITMENT_ROUND_NOT_OPEN.getMessage());
         }
